Rename WooPosVariableProductNavigationData to WooPosVariationsNavigationData and update references across the codebase
