+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/home/items/variations/WooPosVariationsNavigationData.kt
@@ -2,7 +2,7 @@ package com.woocommerce.android.ui.woopos.home.items.variations
 
 import com.woocommerce.android.ui.woopos.util.analytics.WooPosAnalyticsEventConstant
 
-data class WooPosVariableProductNavigationData(
+data class WooPosVariationsNavigationData(
     val id: Long,
     val name: String,
     val numOfVariations: Int,
